# [Optimization] use a pool to reuse LogicalTokenBlock.token_ids