It is surprisingly effective.

Before:

Throughput: 32.37 requests/s, 16574.05 tokens/s

After:

Throughput: 35.64 requests/s, 18250.05 tokens/s

benchmark command: `python benchmarks/benchmark_throughput.py --output-len 256 --input 256 --model meta-llama/Llama-2-7b-hf -tp 8`

machine: 8*H100