@youkaichao 2024-06-17
I also benchmarked it using sharegpt:

`python benchmarks/benchmark_throughput.py --output-len 256 --dataset ShareGPT_V3_unfiltered_cleaned_split.json --model meta-llama/Llama-2-7b-hf -tp 8`

before:

Throughput: 30.89 requests/s, 15854.23 tokens/s

after:

Throughput: 33.84 requests/s, 17369.35 tokens/s

I don't know why the performance benchmark does not show any benefit.

cc @KuntaiDu