@youkaichao 2024-06-17
script: `python benchmarks/benchmark_throughput.py --output-len 256 --dataset ShareGPT_V3_unfiltered_cleaned_split.json --model meta-llama/Llama-2-7b-hf -tp 1`

# Using a single H100

before:

Throughput: 18.43 requests/s, 9461.38 tokens/s

After:

Throughput: 18.49 requests/s, 9491.33 tokens/s

# Using A100 * 4:

before:

Throughput: 16.84 requests/s, 8642.64 tokens/s

After:

Throughput: 16.58 requests/s, 8510.95 tokens/s

# Using a single A100:

before:

Throughput: 9.30 requests/s, 4773.91 tokens/s

after:

Throughput: 9.32 requests/s, 4786.32 tokens/s