@youkaichao 2024-06-17
> might be interesting to see the improvement for 70B model with 4x H100 too...

@njhill updated at https://github.com/vllm-project/vllm/pull/5584#issuecomment-2172218557 .  Also see 5% performance gain.